+++ src/cli_registry.py
@@ -1,0 +1,39 @@
+"""CLI Registry utilities."""
+
+import logging
+import sys
+
+from constants import ExitCodes, PackageManagers
+
+
+def scan_source(pkgtype, dir_name, recursive=False):
+    """Scans the source directory for packages."""
+    if pkgtype == PackageManagers.NPM.value:
+        from registry import npm as _npm  # pylint: disable=import-outside-toplevel
+        return _npm.scan_source(dir_name, recursive)
+    if pkgtype == PackageManagers.MAVEN.value:
+        from registry import maven as _maven  # pylint: disable=import-outside-toplevel
+        return _maven.scan_source(dir_name, recursive)
+    if pkgtype == PackageManagers.PYPI.value:
+        from registry import pypi as _pypi  # pylint: disable=import-outside-toplevel
+        return _pypi.scan_source(dir_name, recursive)
+    logging.error("Selected package type doesn't support import scan.")
+    sys.exit(ExitCodes.FILE_ERROR.value)
+
+
+def check_against(check_type, _level, check_list):
+    """Checks the packages against the registry."""
+    if check_type == PackageManagers.NPM.value:
+        # Fetch details for all levels (fix regression where repo fields were empty on compare)
+        should_fetch_details = True
+        from registry import npm as _npm  # pylint: disable=import-outside-toplevel
+        _npm.recv_pkg_info(check_list, should_fetch_details)
+    elif check_type == PackageManagers.MAVEN.value:
+        from registry import maven as _maven  # pylint: disable=import-outside-toplevel
+        _maven.recv_pkg_info(check_list)
+    elif check_type == PackageManagers.PYPI.value:
+        from registry import pypi as _pypi  # pylint: disable=import-outside-toplevel
+        _pypi.recv_pkg_info(check_list)
+    else:
+        logging.error("Selected package type doesn't support registry check.")
+        sys.exit(ExitCodes.FILE_ERROR.value)
