+++ src/depgate.egg-info/SOURCES.txt
@@ -4,13 +4,28 @@ NOTICE
 README.md
 pyproject.toml
 src/args.py
+src/cli_build.py
+src/cli_classify.py
+src/cli_config.py
+src/cli_io.py
+src/cli_registry.py
 src/constants.py
 src/depgate.py
 src/metapackage.py
 src/analysis/__init__.py
+src/analysis/analysis_runner.py
+src/analysis/facts.py
 src/analysis/heuristics.py
+src/analysis/policy.py
+src/analysis/policy_comparators.py
+src/analysis/policy_rules.py
+src/analysis/policy_runner.py
 src/common/__init__.py
 src/common/http_client.py
+src/common/http_errors.py
+src/common/http_metrics.py
+src/common/http_policy.py
+src/common/http_rate_middleware.py
 src/common/logging_utils.py
 src/depgate.egg-info/PKG-INFO
 src/depgate.egg-info/SOURCES.txt
@@ -19,6 +34,8 @@ src/depgate.egg-info/entry_points.txt
 src/depgate.egg-info/requires.txt
 src/depgate.egg-info/top_level.txt
 src/registry/__init__.py
+src/registry/depsdev/client.py
+src/registry/depsdev/enrich.py
 src/registry/maven/__init__.py
 src/registry/maven/client.py
 src/registry/maven/discovery.py
@@ -36,6 +53,7 @@ src/registry/pypi/scan.py
 src/repository/__init__.py
 src/repository/github.py
 src/repository/gitlab.py
+src/repository/license_discovery.py
 src/repository/provider_adapters.py
 src/repository/provider_registry.py
 src/repository/provider_validation.py
@@ -55,18 +73,31 @@ src/versioning/resolvers/maven.py
 src/versioning/resolvers/npm.py
 src/versioning/resolvers/pypi.py
 tests/test_client_logging.py
+tests/test_depsdev_client_unit.py
+tests/test_depsdev_enrich_unit.py
+tests/test_depsdev_enrich_unit_maven.py
 tests/test_discovery_enrichment_logging.py
 tests/test_github_client.py
 tests/test_gitlab_client.py
 tests/test_heuristics_breakdown_logging.py
 tests/test_heuristics_repo_signals.py
+tests/test_http_client_wrapped_unit.py
+tests/test_http_metrics_unit.py
+tests/test_http_policy_unit.py
+tests/test_license_discovery.py
 tests/test_logging_integration_e2e.py
 tests/test_logging_utils_formatters.py
 tests/test_logging_utils_redaction.py
 tests/test_maven_repo_discovery.py
+tests/test_npm_exists_preservation.py
 tests/test_npm_repo_discovery.py
 tests/test_parse_tokens.py
+tests/test_policy_comparators.py
+tests/test_policy_engine_integration.py
+tests/test_policy_evaluators.py
+tests/test_policy_example_yaml.py
 tests/test_provider_validation_matching.py
+tests/test_pypi_license_enrichment.py
 tests/test_pypi_repo_discovery.py
 tests/test_repo_url_normalize.py
 tests/test_resolver_maven.py
@@ -75,4 +106,5 @@ tests/test_resolver_pypi.py
 tests/test_rtd.py
 tests/test_score_normalization.py
 tests/test_serialization_exports.py
+tests/test_serialization_policy_outputs.py
 tests/test_version_match.py
